refactor: move splitText
